feat: 04 exercise complete
This one was the hardest one yet, I had to use a lot of different solutions before I found one that did what I wanted it to do. In the end I ended up putting two flex boxes on top of each other and aligning the items. Although it got the job done it dont think it was the most efficient method, so I will have to learn how to make a flexbox grid. Apart from that I also need to figure out how to make a max width for a website so that everything doesnt go out of place when you collapse it too much.