feat: 05 exercise completed
Worked out a method to have multiple items on a single row, that involves playing around with the flex basis. After looking at the solution post completion I realize that this is not the most efficient way to solve this problem, but I am yet to completely understand what they did in the solution.