feat: 07 exercise completed
I completed the exercise and added my own touch to a couple things to play around with the transition, transform and hover functions. Overall the exercise was easy to complete, I just had to line everything uup first then commit the changes to text etc. I could have spent more time focusing on the functionality of how the page operates when growing and shrinking the browser.